// WorkSpace/projects/user/src/app/models/Project/iproject-read.ts
export interface IProjectRead {
  id: number;
  name: string;
  description: string;
  address: string;
  image: string;
  siteEngineerName: string;
  projectProgressPercentage: number;
  companyId: number;
  companyName: string;
  companySlogan: string;
  companyEmployees: number;
  companyImage: string;
  companyEmail: string;
  companyPhone: string;
  companyLink: string;
}